Renaming plugin directory
