+++ lib/plugins/bay_opt/bloc/bayesian_optimization_bloc.dart
@@ -1,11 +1,11 @@
 import 'package:bio_flutter/bio_flutter.dart';
-import 'package:biocentral/plugins/bayesian-optimization/bloc/bayesian_optimization_commands.dart';
-import 'package:biocentral/plugins/bayesian-optimization/data/bayesian_optimization_client.dart';
-import 'package:biocentral/plugins/bayesian-optimization/domain/bayesian_optimization_repository.dart';
-import 'package:biocentral/plugins/bayesian-optimization/model/bayesian_optimization_model_types.dart';
-import 'package:biocentral/plugins/bayesian-optimization/model/bayesian_optimization_training_result.dart';
-import 'package:biocentral/plugins/bayesian-optimization/presentation/dialogs/bayesian_optimization_training_dialog_bloc.dart';
-import 'package:biocentral/plugins/bayesian-optimization/presentation/dialogs/start_bayesian_optimization_dialog.dart';
+import 'package:biocentral/plugins/bay_opt/bloc/bayesian_optimization_commands.dart';
+import 'package:biocentral/plugins/bay_opt/data/bayesian_optimization_client.dart';
+import 'package:biocentral/plugins/bay_opt/domain/bayesian_optimization_repository.dart';
+import 'package:biocentral/plugins/bay_opt/model/bayesian_optimization_model_types.dart';
+import 'package:biocentral/plugins/bay_opt/model/bayesian_optimization_training_result.dart';
+import 'package:biocentral/plugins/bay_opt/presentation/dialogs/bayesian_optimization_training_dialog_bloc.dart';
+import 'package:biocentral/plugins/bay_opt/presentation/dialogs/start_bayesian_optimization_dialog.dart';
 import 'package:biocentral/plugins/embeddings/data/predefined_embedders.dart';
 import 'package:biocentral/sdk/biocentral_sdk.dart';
 import 'package:event_bus/event_bus.dart';
