+++ lib/plugins/bay_opt/data/bay_opt_dto.dart
