+++ lib/plugins/bay_opt/model/bayesian_optimization_model_types.dart
