+++ lib/plugins/bay_opt/model/bayesian_optimization_training_result.dart
