+++ lib/plugins/bay_opt/presentation/views/bayesian_optimization_command_view.dart
@@ -1,8 +1,8 @@
-import 'package:biocentral/plugins/bayesian-optimization/bloc/bayesian_optimization_bloc.dart';
-import 'package:biocentral/plugins/bayesian-optimization/model/bayesian_optimization_model_types.dart';
-import 'package:biocentral/plugins/bayesian-optimization/presentation/dialogs/bayesian_optimization_training_dialog_bloc.dart';
-import 'package:biocentral/plugins/bayesian-optimization/presentation/dialogs/iterate_training_dialog.dart';
-import 'package:biocentral/plugins/bayesian-optimization/presentation/dialogs/start_bayesian_optimization_dialog.dart';
+import 'package:biocentral/plugins/bay_opt/bloc/bayesian_optimization_bloc.dart';
+import 'package:biocentral/plugins/bay_opt/model/bayesian_optimization_model_types.dart';
+import 'package:biocentral/plugins/bay_opt/presentation/dialogs/bayesian_optimization_training_dialog_bloc.dart';
+import 'package:biocentral/plugins/bay_opt/presentation/dialogs/iterate_training_dialog.dart';
+import 'package:biocentral/plugins/bay_opt/presentation/dialogs/start_bayesian_optimization_dialog.dart';
 import 'package:biocentral/plugins/embeddings/data/predefined_embedders.dart';
 import 'package:biocentral/sdk/biocentral_sdk.dart';
 import 'package:flutter/material.dart';
