+++ lib/plugins/bay_opt/presentation/views/bayesian_optimization_plot_view.dart
@@ -1,4 +1,4 @@
-import 'package:biocentral/plugins/bayesian-optimization/model/bayesian_optimization_training_result.dart';
+import 'package:biocentral/plugins/bay_opt/model/bayesian_optimization_training_result.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
